pyproject.toml
[tool.poetry]
name = "lens-api"
version = "0.1.0"
description = ""
authors = ["YoungWoongPark <assayw119@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = "^0.111.0"
mangum = "^0.17.0"
SQLAlchemy = "^2.0.30"
pymysql = "^1.1.0"
pydantic = "^2.7.1"
pydantic-core = "^2.18.2"
uvicorn = "^0.29.0"
python-jose = "^3.3.0"
pydantic-settings = "^2.2.1"
sqlmodel = "^0.0.18"
langchain = "^0.2.5"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.5"
aiosqlite = "^0.20.0"
databases = "^0.9.0"
mysql = "^0.0.3"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
